Extract tests
Extract wiki tests into separate file.
